+++ gump/GumpSkills.cpp
@@ -320,7 +320,7 @@ void TGumpSkills::GenerateFrame(int posX, int posY)
 						if (selname == 2) //Изменение имени группы
 						{
 							//Отобразим цветной фон
-							g_GL.DrawPolygone(selectedPolygoneColor, (float)drawX + 16.0f, (float)drawY, 200.0f, 14.0f);
+							g_GL.DrawPolygone(selectedPolygoneColor, drawX + 16, drawY, 200, 14);
 
 							//Текущее название
 							TextEntry->DrawA(6, 0, drawX + 16, drawY - 5);
@@ -334,9 +334,9 @@ void TGumpSkills::GenerateFrame(int posX, int posY)
 
 							//Если выбрана группа или перетаскиваем скилл на текущую группу - подсветим ее
 							if (selname == 1)
-								g_GL.DrawPolygone(selectedPolygoneColor, drawX + 16.0f, (float)drawY, th.Width - 10.0f, 14.0f);
+								g_GL.DrawPolygone(selectedPolygoneColor, drawX + 16, drawY, th.Width - 10, 14);
 							else if (hightlightGroup)
-								g_GL.DrawPolygone(selectedPolygoneColor, drawX + 16.0f, (float)drawY, 200.0f, 14.0f);
+								g_GL.DrawPolygone(selectedPolygoneColor, drawX + 16, drawY, 200, 14);
 
 							//Отобразим имя группы (из заранее заготовленной текстуры)
 							th.Draw(drawX + 16, drawY - 5);
@@ -404,7 +404,7 @@ void TGumpSkills::GenerateFrame(int posX, int posY)
 									if (!th.Empty())
 									{
 										if (si == idx)
-											g_GL.DrawPolygone(selectedPolygoneColor, drawX + 20.0f, drawY - 1.0f, 250.0f, 14.0f);
+											g_GL.DrawPolygone(selectedPolygoneColor, drawX + 20, drawY - 1, 250, 14);
 
 										th.Draw(drawX + 22, drawY - 1);
 									}
